Modifying for phase 3
