+++ phase_3/TestHarness/ITest.h
@@ -1,0 +1,10 @@
+#include <string>
+
+class ITest
+{
+public:
+	// Implements to run test(s)
+	virtual bool execute() = 0;
+
+	bool runTest(bool (*function)());
+};
