+++ phase_3/TestHarness/Software.h
@@ -1,0 +1,3 @@
+int testFunction1();
+int testFunction2();
+int testFunction3();
