+++ phase_3/TestHarness/TestHarness.vcxproj
@@ -64,13 +64,16 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClInclude Include="ITest.h" />
     <ClInclude Include="Logger.h" />
+    <ClInclude Include="Software.h" />
     <ClInclude Include="TestEngine.h" />
     <ClInclude Include="UnitTest.h" />
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="Logger.cpp" />
     <ClCompile Include="Main.cpp" />
+    <ClCompile Include="Software.cpp" />
     <ClCompile Include="TestEngine.cpp" />
     <ClCompile Include="UnitTest.cpp" />
   </ItemGroup>
