+++ phase_3/TestHarness/UnitTest.cpp
@@ -1,7 +1,5 @@
 #include "UnitTest.h"
 
-UnitTest::UnitTest() {}
-
 // the assertion stuct is passed in to the assert function
 void UnitTest::AssertEquals(
     int (*function)(),
@@ -26,21 +24,3 @@ vector<assertion> UnitTest::GetAssertions()
 }
 */
 
-int testFunction1()
-{
-    // test passed
-    return 0;
-}
-
-int testFunction2()
-{
-    // test failed
-    return -1;
-}
-
-int testFunction3()
-{
-    // exception raised
-    std::string("abc").substr(10);
-    return 0;
-}
