+++ phase_3/TestHarness/UnitTest.h
@@ -10,10 +10,6 @@ using std::vector;
 using std::string;
 
 // callable object that does not take any parameters
-int testFunction1();
-int testFunction2();
-int testFunction3();
-
 struct assertion {
 	string assertion_type = "test";
 	string function_name = "func";
@@ -30,15 +26,11 @@ private:
 
 public:
 
-	// default constructor
-	UnitTest();
+	// Test method to override
+	virtual bool execute() = 0;
 
 	// assertion - exeuction function
-	void AssertEquals(
-		int (*function)(), 
-		int test, 
-		assertion assert
-	);
+	static void AssertEquals(int (*function)(), int test, assertion assert);
 
 	// getter for retrieving tests and their results
 	// vector<assertion> GetAssertions();
